django-app/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
idna==2.9
praw==7.0.0
prawcore==1.4.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
update-checker==0.17
urllib3==1.25.9
websocket-client==0.57.0
whitenoise==5.1.0
gunicorn
dj-database-url==0.5.0